Refactor build command into Project class
